# pkg/api/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Workflow management module",
        "title": "CM-Cicada REST API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "latest"
    },
    "basePath": "/cicada",
    "paths": {
...
        "/task/{taskId}": {
            "get": {
                "description": "Get the task directly.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Get Task Directly",
                "operationId": "get-task-directly",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the task.",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get the task.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.TaskDirectly"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the task.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workflow/{wfId}/task": {
            "get": {
                "description": "Get a task list of the workflow.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "List Task",
                "operationId": "list-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get a task list.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Task"
                            }
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get a task list.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workflow/{wfId}/task/{taskId}": {
            "get": {
                "description": "Get the task.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Get Task",
                "operationId": "get-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the task.",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get the task.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Task"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the task.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workflow/{wfId}/task_group/{tgId}/task": {
            "get": {
                "description": "Get a task list from the task group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "List Task from Task Group",
                "operationId": "list-task-from-task-group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the task group.",
                        "name": "tgId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get a task list from the task group.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Task"
                            }
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get a task list from the task group.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workflow/{wfId}/task_group/{tgId}/task/{taskId}": {
            "get": {
                "description": "Get the task from the task group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Get Task from Task Group",
                "operationId": "get-task-from-task-group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the task group.",
                        "name": "tgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the task.",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get the task from the task group.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Task"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the task from the task group.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Task": {
            "type": "object",
            "required": [
                "id",
                "name",
                "request_body",
                "task_component"
            ],
            "properties": {
                "dependencies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "is_deleted_task": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "path_params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "query_params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "request_body": {
                    "type": "string"
                },
                "task_component": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.TaskDirectly": {
            "type": "object",
            "required": [
                "id",
                "name",
                "request_body",
                "task_component",
                "task_group_id",
                "workflow_id"
            ],
            "properties": {
                "dependencies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path_params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "query_params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "request_body": {
                    "type": "string"
                },
                "task_component": {
                    "type": "string"
                },
                "task_group_id": {
                    "type": "string"
                },
                "workflow_id": {
                    "type": "string"
                }
            }
        },
...
    }
}